custom tool creation is now working and the test screen as well
