+++ admin/assets/javascripts/discourse/templates/admin-plugins/show/discourse-ai-tools/show.hbs
@@ -1,0 +1,1 @@
+<AiToolEditor @tools={{this.allTools}} @model={{this.model}} />
