+++ app/controllers/discourse_ai/admin/ai_tools_controller.rb
@@ -78,7 +78,6 @@ module DiscourseAi
           :description,
           :script,
           :summary,
-          :details,
           parameters: %i[name type description],
         )
       end
