+++ app/serializers/ai_custom_tool_serializer.rb
@@ -4,6 +4,7 @@ class AiCustomToolSerializer < ApplicationSerializer
   attributes :id,
              :name,
              :description,
+             :summary,
              :parameters,
              :script,
              :created_by_id,
