+++ config/locales/client.en.yml
@@ -187,7 +187,11 @@ en:
         short_title: "Tools"
         new: "New Tool"
         name: "Name"
+        name_help: "The unique name of the tool as used by the language model"
         description: "Description"
+        description_help: "A clear description of the tool's purpose for the language model"
+        summary: "Summary"
+        summary_help: "Summary of tools purpose to be displayed to end users"
         script: "Script"
         parameters: "Parameters"
         save: "Save"
@@ -197,6 +201,13 @@ en:
         parameter_name: "Parameter Name"
         parameter_description: "Parameter Description"
         edit: "Edit"
+        test: "Run Test"
+        delete: "Delete"
+        saved: "Tool saved"
+        test_modal:
+          title: "Test AI Tool"
+          run: "Run Test"
+          result: "Test Result"
 
       llms:
         short_title: "LLMs"
