+++ spec/models/ai_tool_spec.rb
@@ -11,7 +11,6 @@ RSpec.describe AiTool do
       parameters: parameters || [{ name: "query", type: "string", desciption: "perform a search" }],
       script: script || "function invoke(params) { return params; }",
       created_by_id: 1,
-      details: "Test tool details",
       summary: "Test tool summary",
     )
   end
